rename `aggregation/state_cache` to `aggregation/minput_agg_impl`
