+++ src/stream/src/executor/aggregation/minput_agg_impl/extreme.rs
