+++ src/stream/src/executor/aggregation/minput_agg_impl/string_agg.rs
