+++ src/stream/src/executor/aggregation/mod.rs
@@ -36,7 +36,7 @@ pub mod agg_impl;
 mod agg_state;
 mod distinct;
 mod minput;
-mod state_cache;
+mod minput_agg_impl;
 mod table;
 mod value;
